checks if the user's  in the queue and logs out nicely

+++ client/app/queue/queue.factory.js
@@ -37,19 +37,20 @@ angular.module('clawFrontApp')
 
         //add player to queue
         factory.addPlayer = function(player) {
-          var alreadyInQueue = false;
-          $rootScope.queue.forEach(function(el){
-              if(el.userId == player._id)
-                  {alreadyInQueue = true;}
-                    return alreadyInQueue;
+            var alreadyInQueue = false;
+            $rootScope.queue.forEach(function(el) {
+                if (el.userId == player._id) {
+                    alreadyInQueue = true;
+                }
+                return alreadyInQueue;
             })
-              if(!alreadyInQueue && player.credits >=1){ 
-              $http.post('/api/queues', {
-                username: player.name,
-                userId: player._id,
-                active: true,
-                index: Date.now()
-              })
+            if (!alreadyInQueue && player.credits >= 1) {
+                $http.post('/api/queues', {
+                    username: player.name,
+                    userId: player._id,
+                    active: true,
+                    index: Date.now()
+                })
             }
         }
 
@@ -57,44 +58,64 @@ angular.module('clawFrontApp')
 
         //remove player from queue
         factory.removeByQueueUserId = function(player) {
-        console.log("player in $http.delete request: ", player)
-            return $http.delete('/api/queues/' + player.userId);
-        }
+            console.log("player in $http.delete request: ", player)
+            var there = findPlayerInQueue($rootScope.queue, player);
+            if (there > 0) {
+                return $http.delete('/api/queues/' + player.userId);
+            } else {
+                return undefined;
+            }
+        };
 
         factory.removeByUserId = function(player) {
-        console.log("player in $http.delete request: ", player)
-            return $http.delete('/api/queues/' + player._id);
-        }
+            console.log("player in $http.delete request: ", player)
+            var there = findPlayerInQueue($rootScope.queue, player);
+            if (there > 0) {
+                return $http.delete('/api/queues/' + player._id);
+            } else {
+                return undefined;
+            }
+        };
+
+        var findPlayerInQueue = function(queue, player) {
+            var isThere = -1;
+            for (var i = 0; i < queue.length; i++) {
+                console.log("queue[i].userId", queue[i].userId, "player", player._id, "i: ", i)
+                if (queue[i].userId == player._id) {
+                    isThere = i;
+                    console.log("found it!", isThere);
+                }
+            }
+            console.log('isThere', isThere);
+            return isThere;
+        };
 
         //check ETAtoPlay
         factory.ETAtoPlay = function(queue, player) {
-           
             var eta = queue.length;
-                console.log("it's running")
+            console.log("it's running")
             for (var i = 0; i < queue.length; i++) {
                 console.log("queue[i].userId", queue[i].userId, "player", player._id, "i: ", i)
                 if (queue[i].userId == player._id) {
                     eta = i;
                     console.log("found it!", eta);
                 }
-               
             }
             console.log("eta", eta)
             return eta;
-        }
+        };
 
         // check if player is in queue
 
         // play game with credits
 
         factory.playWithCredits = function(player) {
-            if (player.credits >=1) player.credits--;
-            $http.put('/api/users/' + user._id, user).success(function() {console.log("$http.put user: ", user);
-      });
+            if (player.credits >= 1) player.credits--;
+            $http.put('/api/users/' + user._id, user).success(function() {
+                console.log("$http.put user: ", user);
+            });
         }
-    
-     
+
+
         return factory;
     });
-
-
